dash/requirements.txt
dash
pandas
plotly.express
dash_core_components
dash_html_components
sqlalchemy 
psycopg2-binary

